third_party/blink/web_tests/http/tests/origin_trials/webexposed/cookiestore-origin-trial-interfaces.html

<!DOCTYPE html>
<meta charset="utf-8">
<!-- Generate token with the command:
generate_token.py http://127.0.0.1:8000 CookieStore --expire-timestamp=2000000000
-- -->
<meta http-equiv="origin-trial" content="AuCNc4F6ez8bdiKV6reoNKgzu2afmtUl5FgKkP6jdrbbCqVh8BfddejNqciWMz+V+oZXxJdW1LU5nQuC0Ij2GQkAAABTeyJvcmlnaW4iOiAiaHR0cDovLzEyNy4wLjAuMTo4MDAwIiwgImZlYXR1cmUiOiAiQ29va2llU3RvcmUiLCAiZXhwaXJ5IjogMjAwMDAwMDAwMH0=" />
<title>Cookie Store API - interfaces exposed by origin trial</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/resources/origin-trials-helper.js"></script>
<script src="/serviceworker/resources/test-helpers.js"></script>
<script>
test(t => {
  OriginTrialsHelper.check_properties_exist(this, {
     'CookieStore': ['get', 'getAll', 'set', 'delete', 'onchange'],
     'CookieChangeEvent': ['changed', 'deleted'],
  });
}, 'Cookie Store API interfaces and properties in Origin-Trial enabled document.');

test(t => {
  assert_true('cookieStore' in self, 'cookieStore property exists on global scope');
}, 'Cookie Store API entry point in Origin-Trial enabled document.');

// Only run "disabled" tests if the feature is not enabled via runtime flags.
if (!self.internals.runtimeFlags.cookieStoreEnabled) {
  service_worker_test('resources/cookiestore-origin-trial-interfaces-serviceworker-disabled.js');
}

service_worker_test('resources/cookiestore-origin-trial-interfaces-serviceworker-enabled.php');

</script>
